Add morphological planform object, begin generalization of mask methods.
This commit creates a preliminary "Morphological Planform" class which
uses morphological operations, namely the binary closing, to transform
a binary input image (expected to be an ElevationMask) into a composite
image similar to the `sea_angles` array returned from the OAM.

The idea is that this image, which we try to standardize as a
"composite_array" which is a property all Planform objects should now
have. This "composite_array" is what we can threshold in the mask
functions.

So in the masking functions, this commit tries to generalize the
methods such that either any BasePlanform is acceptable in some
circumstances, or a "method" argument can be provided to specify the
type of Planform to compute (OAM or MPM). This is incomplete as
of this commit.