revise input checking for mobility functions
This commit revises the input checking for the mobility functions.
Now the input checker can digest time-coordinate inputs or
time-index values, and it spits out dimensioned xarray objects
as well as converted (if needed) indices for the time values.

Revises the testing to get things working. Still need to do
the doc-strings and change the actual mobility functions to
stop returning arrays and return dimensioned xarrays...